Foundation, Catalog/Job Query setup, & Job Process Test,
The core concept of the program, involving grabbing MFG files in a folder structure and populating information, is mostly done but needs to actually implemented via a function or whatever.
The Assembled/Ship-loose mess of classes that were the downfall of maintainability in the predecessor of this, as it got over complicated with not fully thinking out how to structure/make it all, is not yet added- but shouldn't be too involved.

Module setup to query the product catalog and get whatever info is needed, as well as query a job's quote products (item lines).
A test was ran using these along with reading the mfg file of the product packages and building an mfg string with populated values; Though, I needs a more concrete implementation than a long one-liner in a new class named "PackageProcessor" or something.