Frames, Package, and Mfg Handlers Created
Frame Handler takes care of parsing the populated mfg lines and creating new instance of classes that represent the hierarchy of groups that are in an mfg. The package mfg hierarchy is: ProductCollection (all assembled frames and ship-loose for product package) -> AssembledFrames (MfgItem for frame) -> MfgItem
MfgItem is the object that is created for every line in the mfg except the partition frame lines. it represents a general mfg line item with properties for each header key and it's value.
PartitionFrame represents every assembled frame. It contains a property for its name, an array of MfgItems that it contains (that are assembled) .
ProductFrameCollection is all of the PartitionFrames (and ship-loose MfgItem objects) that were all processed for a product mfg package. It has two arrays for both the assembled frames objects and another for the ship-loose, as well as a property to store the name of the name of the product that the package was for.

PackageHandler reads the package mfg, populating the placeholders, and using the FrameHandler to process it. Stores the result of processing the packages in ProductCollection array.

MfgHandlers converts the class objects back to formatted mfg lines and saves that result to a file. Most of the work is taken care of by the toString() function on each of the class objects, the MfgHandler just joins together the ProductCollection (which calls toString() on each Frame, which then does it to each MfgItem in the frame, and the ship-loose MfgItem objects) objects' mfg snippets and the mfg header.

Misc:
PartitionFrame modle file created, contains all the class objects mentioned above with the FrameHandler.
Catalog query result gets cached so it doesn't take over 5 seconds just to fetch the catalog every time. Deleting the cached json file will have it re-query the data and save it to the cache file.
Exports and mfg file with extra newlines as spacers between each partition frame, ship-loose items, and other package mfg contents for now to test.
General Tidying up, though could use more polishing specifically in the documenting/commenting department.
No support for handling the Frame Wks (1" tops) when it comes to recognizing grommets/dean-z's ordered on the same quote part and providing the new generated product code for that variant. Might take a little bit to get this converted considering this section of code in the original program is the messiest part of the code D: .