Frame ShipLoose Bug, Converting CRM/Dyn Item Types, & Misc
INI files added for use (described later) of translating between CRM and Dynamics item types, as they are different.
FrameHandler modified to fix a bug where it wouldn't include lines that contained the package name. This was meant to not include frame lines in the ship-loose but also excluded mfg one-liner packages from their own mfg text.
PackageHandler slightly rewritten to filter out quote line items that don't exist in the product catalog (optional). Also is smarter at getting the correct package path, either system generated or user-created- and filtering out lines that don't match the criteria for mfg lines.
ItemCatalog now has a function, and dictionary that is populated with the same data, that returns an object containing a string for the equivalent of a provided type forr both CRM and Dyn. Translating the types is required to find packages in the Type/SubType folder structure and then also then the MfgItems are created in FrameHandler to ensure CRM will process them correctly.
Miscellaneous tidying up.