MFG Export Includes Contextual Info Now
The MFG export will now separate processed snippets based on the package that they came from. The separators will be empty MFG lines that are populated with the package name (and qty) and the last time the package's MFG file was modified (at the time of generating). These line(s) are just to provide more contextual information within the generated file if it would every be useful in the future.
These added Mfg lines don't have anything in the product code column so they are ignored when importing into CRM and just function as comments/notes.